intermediate up for large files: long long int

+++ code/uhttpd-0.0.2/src/response.c
@@ -56,7 +56,7 @@ int  doresponse(int sd){
 	char        *p1, *p2;
 
 	int         status_code;
-	int         bytes_sent = 0;
+	long long int    bytes_sent = 0;
 
 	int         ret = 0;
